requirements.txt
fastapi==0.111.0
python_dotenv==1.0.1
psycopg-binary==3.1.18
psycopg==3.1.18
